builtins::fs_meta

Function fs_crawl

fn fs_crawl(
    environment: &mut SloshVm,
    path: String,
    lambda_exp: Value,
    optional_depth_or_symlink: VarArgs<Value>,
) -> VMResult<Value>

Usage: (fs-crawl /path/to/file/or/dir (fn (x) (prn “found path” x) [max-depth] [:follow-syms])

If a directory is provided the path is recursively searched and every file and directory is called as an argument to the provided function. If a file is provided the path is provided as an argument to the provided function. Takes two optional arguments (in any order) an integer, representing max depth to traverse if file is a directory, or the symbol, :follow-syms, to follow symbol links when traversing if desired.

Section: file

Example:

(with-temp-file (fn (tmp-file) (let (cnt 0) (fs-crawl tmp-file (fn (x) (test::assert-equal (fs-base tmp-file) (fs-base x)) (set! cnt (+ 1 cnt)))) (test::assert-equal 1 cnt))))

(defn create-in (in-dir num-files visited) (dotimes-i i num-files (let (tmp-file (get-temp-file in-dir)) (set! visited.~tmp-file #f))))

(defn create-dir (tmp-dir visited) (let (new-tmp (get-temp tmp-dir)) (set! visited.~new-tmp #f) new-tmp))

(with-temp (fn (root-tmp-dir) (let (tmp-file-count 5 visited {} cnt 0) (set! visited.~root-tmp-dir #f) (create-in root-tmp-dir tmp-file-count visited) (let (tmp-dir (create-dir root-tmp-dir visited) new-files (create-in tmp-dir tmp-file-count visited) tmp-dir (create-dir tmp-dir visited) new-files (create-in tmp-dir tmp-file-count visited)) (fs-crawl root-tmp-dir (fn (x) (let (file visited.~x) (test::assert-true (not file)) ;; also tests double counting (set! visited.~x #t) (inc! cnt)))) (test::assert-equal (+ 3 (* 3 tmp-file-count)) cnt) (test::assert-equal (+ 3 (* 3 tmp-file-count)) (len visited)) (seq-for key in (hash-keys visited) (test::assert-true visited.~key))))))

(with-temp (fn (root-tmp-dir) (let (tmp-file-count 5 visited {} cnt 0) (set! visited.~root-tmp-dir #f) (create-in root-tmp-dir tmp-file-count visited) (let (tmp-dir (create-dir root-tmp-dir visited) new-files (create-in tmp-dir tmp-file-count visited) tmp-dir (create-dir tmp-dir {}) new-files (do (set! visited.~tmp-dir #f)(create-in tmp-dir tmp-file-count {}))) (fs-crawl root-tmp-dir (fn (x) (let (file visited.~x) (test::assert-true (not file)) ;; also tests double counting (set! visited.~x #t) (inc! cnt))) 2) (test::assert-equal (+ 3 (* 2 tmp-file-count)) cnt) (test::assert-equal (+ 3 (* 2 tmp-file-count)) (len visited)) (seq-for key in (hash-keys visited) (test::assert-true visited.~key))))))

(with-temp (fn (root-tmp-dir) (let (tmp-file-count 5 visited {} cnt 0) (set! visited.~root-tmp-dir #f) (create-in root-tmp-dir tmp-file-count visited) (let (tmp-dir (create-dir root-tmp-dir {}) new-files (do (set! visited.~tmp-dir #f)(create-in tmp-dir tmp-file-count {})) tmp-dir (create-dir tmp-dir {}) new-files (create-in tmp-dir tmp-file-count {})) (fs-crawl root-tmp-dir (fn (x) (let (file visited.~x) (test::assert-true (not file)) ;; also tests double counting (set! visited.~x #t) (inc! cnt))) 1) (test::assert-equal (+ 2 tmp-file-count) cnt) (test::assert-equal (+ 2 tmp-file-count) (len visited)) (seq-for key in (hash-keys visited) (test::assert-true visited.~key))))))