builtins::string

Function parse_to_int

fn parse_to_int(environment: &mut SloshVm, args: &[Value]) -> VMResult<Value>

Usage: (->int ?int) -> int

If string or other value is a valid representation of an integer return that int. Error if not.

Note, if value is a float (or is a string that casts initially to a float) and it is NaN, +/- infinity, < i56::MIN, or > i56::MAX the function will error.

Section: type

Example: (test::assert-equal 0 (->int “0”)) (test::assert-equal 101 (->int “101”)) (test::assert-equal 1101 (->int :1101)) (test::assert-equal 8 (->int 8.1)) (test::assert-equal -111 (->int “-111”)) (test::assert-equal 10 (->int “10.0”)) (test::assert-error (->int “not int”)) (test::assert-error (->int “–10”)) ;; for some reason this ->sym test case must be last or else there’s an error w/ load (test::assert-equal 1102 (->int (->sym 1102)))